Add Docker test environment for easy development.
Add a Dockerfile with a Postgres build environment as well as simple
scripting to make it easy for contributors to quickly iterate on
extension development without needing to configure a build environment.

Unfortunately I can't easily add the Docker setup to the Makefile (as
it's normally common to do) since the existing Makefile requires the
full Postgres extension build environment, which would defeat a lot of
the purpose of the Docker environment in the first place.